Add dashboard wide social icons for support
