+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/index.tsx
@@ -5,6 +5,7 @@ import { withTheme } from 'styled-components';
 
 import { Chat } from './Chat';
 import { Container } from './elements';
+import { Help } from './socials';
 import { Comments } from './screens/Comments';
 import { ConfigurationFiles } from './screens/ConfigurationFiles';
 import { Deployment } from './screens/Deployment/index';
@@ -64,7 +65,7 @@ export const WorkspaceComponent = ({ theme }) => {
             {editor.currentSandbox && <Component />}
           </div>
           {isLive && roomInfo.chatEnabled && <Chat />}
-          Need help? ICons
+          <Help />
         </>
       </ThemeProvider>
     </Container>
