+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/screens/ProjectInfo/Delete.tsx
@@ -1,22 +1,8 @@
-import { Button, Stack, Element, Link } from '@codesandbox/components';
+import { Button, Stack } from '@codesandbox/components';
 import css from '@styled-system/css';
 import React, { FunctionComponent, MouseEvent } from 'react';
 
 import { useAppState, useActions } from 'app/overmind';
-
-import { DiscordIcon, GithubIcon, TwitterIcon } from './social-icons';
-
-const links = [
-  { href: 'https://twitter.com/codesandbox', Icon: TwitterIcon },
-  {
-    href: 'https://github.com/codesandbox/codesandbox-client',
-    Icon: GithubIcon,
-  },
-  {
-    href: 'https://discord.gg/C6vfhW3H6e',
-    Icon: DiscordIcon,
-  },
-];
 
 export const Delete: FunctionComponent = () => {
   const {
@@ -38,37 +24,19 @@ export const Delete: FunctionComponent = () => {
   };
 
   return (
-    <Element>
-      <Stack justify="center" marginBottom={6}>
-        <Button
-          css={css({
-            ':hover:not(:disabled),:focus:not(:disabled)': {
-              color: 'errorForeground',
-            },
-          })}
-          // @ts-ignore
-          onClick={onDelete}
-          variant="link"
-        >
-          {`Delete ${customTemplate ? `Template` : `Sandbox`}`}
-        </Button>
-      </Stack>
-
-      <Element marginBottom={3} paddingX={2}>
-        <Stack>
-          <p>Need help? ICons</p>
-          {links.map(({ href, Icon }) => (
-            <Link
-              css={css({ color: 'sideBar.border' })}
-              href={href}
-              rel="noopener noreferrer"
-              target="_blank"
-            >
-              <Icon />
-            </Link>
-          ))}
-        </Stack>
-      </Element>
-    </Element>
+    <Stack justify="center" marginBottom={6}>
+      <Button
+        css={css({
+          ':hover:not(:disabled),:focus:not(:disabled)': {
+            color: 'errorForeground',
+          },
+        })}
+        // @ts-ignore
+        onClick={onDelete}
+        variant="link"
+      >
+        {`Delete ${customTemplate ? `Template` : `Sandbox`}`}
+      </Button>
+    </Stack>
   );
 };
